+++ packages/google-pay-integration/src/gateways/google-pay-gateway.ts
@@ -32,6 +32,8 @@ import itemsRequireShipping from '../utils/items-require-shipping';
 
 export default class GooglePayGateway {
     private _getPaymentMethodFn?: () => PaymentMethod<GooglePayInitializationData>;
+    private _isBuyNowFlow = false;
+    private _currencyCode?: string;
 
     constructor(
         private _gatewayIdentifier: string,
@@ -85,11 +87,24 @@ export default class GooglePayGateway {
         return Promise.resolve({
             nonce,
             card_information: { type, number },
+            ...(this._isBuyNowFlow && {
+                cart_id: this._paymentIntegrationService.getState().getCart()?.id,
+            }),
         });
     }
 
     async getRequiredData(): Promise<GooglePayRequiredPaymentData> {
         const data: GooglePayRequiredPaymentData = { emailRequired: true };
+
+        if (this._isBuyNowFlow) {
+            return {
+                ...data,
+                shippingAddressRequired: true,
+                shippingAddressParameters: {
+                    phoneNumberRequired: true,
+                },
+            };
+        }
 
         if (this._isShippingAddressRequired()) {
             const state = await this._paymentIntegrationService.loadShippingCountries();
@@ -135,6 +150,14 @@ export default class GooglePayGateway {
     }
 
     getTransactionInfo(): GooglePayTransactionInfo {
+        if (this._isBuyNowFlow) {
+            return {
+                currencyCode: this._getCurrencyCodeOrThrow(),
+                totalPriceStatus: TotalPriceStatusType.ESTIMATED,
+                totalPrice: '0',
+            };
+        }
+
         const { getCheckoutOrThrow, getCartOrThrow } = this._paymentIntegrationService.getState();
         const countryCode = this.getGooglePayInitializationData().storeCountry;
         const { code: currencyCode, decimalPlaces } = getCartOrThrow().currency;
@@ -179,8 +202,18 @@ export default class GooglePayGateway {
         };
     }
 
-    initialize(getPaymentMethod: () => PaymentMethod<GooglePayInitializationData>): Promise<void> {
+    initialize(
+        getPaymentMethod: () => PaymentMethod<GooglePayInitializationData>,
+        isBuyNowFlow?: boolean,
+        currencyCode?: string,
+    ): Promise<void> {
         this._getPaymentMethodFn = getPaymentMethod;
+        this._isBuyNowFlow = Boolean(isBuyNowFlow);
+        this._currencyCode = currencyCode;
+
+        if (this._isBuyNowFlow) {
+            this._getCurrencyCodeOrThrow();
+        }
 
         return Promise.resolve();
     }
@@ -258,5 +291,15 @@ export default class GooglePayGateway {
         const lastName = nameParts[nameParts.length - 1];
 
         return [firstName, lastName];
+    }
+
+    private _getCurrencyCodeOrThrow(): string {
+        return guard(
+            this._currencyCode,
+            () =>
+                new InvalidArgumentError(
+                    'Unable to initialize payment because "options.currencyCode" argument is not provided.',
+                ),
+        );
     }
 }